fix(web): keep pg out of automations client bundle
Split client-safe automation schemas and types from the Drizzle store so the new-automation page no longer pulls node-postgres into the browser build.
