+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/automations/_components/automation-msw-handlers.ts
@@ -12,7 +12,7 @@
  */
 import { http, HttpResponse } from "msw";
 
-import type { WorkspaceAutomationRecord } from "@/lib/agents/workspace-automations";
+import type { WorkspaceAutomationRecord } from "@/lib/agents/workspace-automation-types";
 
 import {
   automationEditorContentfulConnectionsFixture,
