+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/dashboard/_components/dashboard-page-content.tsx
@@ -27,7 +27,7 @@ import { mapWithConcurrency } from "@/lib/primitives/map-with-concurrency/map-wi
 import { isNativeWorkspaceJob } from "@/lib/projects/workspace-resource-capabilities";
 import { getTmsProviderBranding } from "@/lib/providers/shared/tms-provider-branding";
 import { readTmsProviderListResponse } from "@/lib/providers/jobs/tms-provider-list-fetch";
-import type { WorkspaceAutomationRunRecord } from "@/lib/agents/workspace-automations";
+import type { WorkspaceAutomationRunRecord } from "@/lib/agents/workspace-automation-types";
 
 import { buildJobDetailHref } from "../../jobs/_components/jobs-view-helpers";
 import {
