+++ apps/hyperlocalise-web/src/lib/agents/workspace-automation-github-mapping.ts
@@ -19,7 +19,7 @@ import type {
   WorkspaceAutomationRecord,
   WorkspaceAutomationToolConfig,
   WorkspaceAutomationTriggerConfig,
-} from "./workspace-automations";
+} from "./workspace-automation-types";
 
 export function resolveWorkspaceAutomationGithubMode(
   toolConfig: WorkspaceAutomationToolConfig,
