+++ apps/hyperlocalise-web/src/lib/agents/workspace-automation-view-model.ts
@@ -17,7 +17,7 @@ import type {
   WorkspaceAutomationToolConfig,
   WorkspaceAutomationTriggerConfig,
   WorkspaceAutomationWebSearchProvider,
-} from "./workspace-automations";
+} from "./workspace-automation-types";
 import { parseSlackConversationId } from "./slack/channel-query";
 import {
   getWorkspaceAutomationTemplate,
